Create text entry as a window port over values
